pyproject.toml
[project]
name = "rag-math-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pylatexenc>=2.10",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "weaviate-client>=4.14.3",
]
